Déplacement des données d'adresse de informationsDeclarant vers informationsEntreprise
